refactor: isolate vault test state handling
